requirements.txt
discord.py-self
termcolor
PyDictionary
python-dotenv
requests
pyowm